split core plugin support from builtin plugins
The definition of plugins is now in squad.core.plugins, while built-in
plugins are in squad.plugins. This allows for example to have plugins
depending on stuff from squad.frontend, without creating a circular
dependency (core → plugins → frontend → core).
